Removed 1.9 Backwards protocols from this project and moved Chunk Types to the correct packages
