+++ common/src/main/java/com/viaversion/viaversion/protocols/protocol1_9_1to1_9/types/Chunk1_9_1_2Type.java
@@ -15,10 +15,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.viaversion.viaversion.protocols.protocol1_9_1_2to1_9_3_4.types;
+package com.viaversion.viaversion.protocols.protocol1_9_1to1_9.types;
 
-import com.github.steveice10.opennbt.tag.builtin.CompoundTag;
-import com.viaversion.viaversion.api.Via;
 import com.viaversion.viaversion.api.minecraft.Environment;
 import com.viaversion.viaversion.api.minecraft.chunks.BaseChunk;
 import com.viaversion.viaversion.api.minecraft.chunks.Chunk;
@@ -30,13 +28,12 @@ import com.viaversion.viaversion.api.type.types.version.Types1_9;
 import com.viaversion.viaversion.protocols.protocol1_9_3to1_9_1_2.storage.ClientWorld;
 import io.netty.buffer.ByteBuf;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.util.BitSet;
 
-public class Chunk1_9_3_4Type extends PartialType<Chunk, ClientWorld> {
+public class Chunk1_9_1_2Type extends PartialType<Chunk, ClientWorld> {
 
-    public Chunk1_9_3_4Type(ClientWorld param) {
-        super(param, Chunk.class);
+    public Chunk1_9_1_2Type(ClientWorld clientWorld) {
+        super(clientWorld, Chunk.class);
     }
 
     @Override
@@ -44,15 +41,23 @@ public class Chunk1_9_3_4Type extends PartialType<Chunk, ClientWorld> {
         int chunkX = input.readInt();
         int chunkZ = input.readInt();
 
-        boolean fullChunk = input.readBoolean();
+        boolean groundUp = input.readBoolean();
         int primaryBitmask = Type.VAR_INT.readPrimitive(input);
+        // Size (unused)
         Type.VAR_INT.readPrimitive(input);
 
-        // Read sections
+        BitSet usedSections = new BitSet(16);
         ChunkSection[] sections = new ChunkSection[16];
+        // Calculate section count from bitmask
         for (int i = 0; i < 16; i++) {
-            if ((primaryBitmask & (1 << i)) == 0) continue; // Section not set
+            if ((primaryBitmask & (1 << i)) != 0) {
+                usedSections.set(i);
+            }
+        }
 
+        // Read sections
+        for (int i = 0; i < 16; i++) {
+            if (!usedSections.get(i)) continue; // Section not set
             ChunkSection section = Types1_9.CHUNK_SECTION.read(input);
             sections[i] = section;
             section.getLight().readBlockLight(input);
@@ -61,24 +66,14 @@ public class Chunk1_9_3_4Type extends PartialType<Chunk, ClientWorld> {
             }
         }
 
-        int[] biomeData = fullChunk ? new int[256] : null;
-        if (fullChunk) {
+        int[] biomeData = groundUp ? new int[256] : null;
+        if (groundUp) {
             for (int i = 0; i < 256; i++) {
                 biomeData[i] = input.readByte() & 0xFF;
             }
         }
 
-        List<CompoundTag> nbtData = new ArrayList<>(Arrays.asList(Type.NBT_ARRAY.read(input)));
-
-        // Read all the remaining bytes (workaround for #681)
-        if (input.readableBytes() > 0) {
-            byte[] array = Type.REMAINING_BYTES.read(input);
-            if (Via.getManager().isDebug()) {
-                Via.getPlatform().getLogger().warning("Found " + array.length + " more bytes than expected while reading the chunk: " + chunkX + "/" + chunkZ);
-            }
-        }
-
-        return new BaseChunk(chunkX, chunkZ, fullChunk, false, primaryBitmask, sections, biomeData, nbtData);
+        return new BaseChunk(chunkX, chunkZ, groundUp, false, primaryBitmask, sections, biomeData, new ArrayList<>());
     }
 
     @Override
@@ -99,6 +94,7 @@ public class Chunk1_9_3_4Type extends PartialType<Chunk, ClientWorld> {
 
                 if (!section.getLight().hasSkyLight()) continue; // No sky light, we're done here.
                 section.getLight().writeSkyLight(buf);
+
             }
             buf.readerIndex(0);
             Type.VAR_INT.writePrimitive(output, buf.readableBytes() + (chunk.isBiomeData() ? 256 : 0));
@@ -113,9 +109,6 @@ public class Chunk1_9_3_4Type extends PartialType<Chunk, ClientWorld> {
                 output.writeByte((byte) biome);
             }
         }
-
-        // Write Block Entities
-        Type.NBT_ARRAY.write(output, chunk.getBlockEntities().toArray(new CompoundTag[0]));
     }
 
     @Override
